feat(merge): add higher-order lettable version of merge (#2809)
