+++ src/operators/index.ts
@@ -35,6 +35,7 @@ export { isEmpty } from './isEmpty';
 export { map } from './map';
 export { materialize } from './materialize';
 export { max } from './max';
+export { merge } from './merge';
 export { mergeAll } from './mergeAll';
 export { mergeMap } from './mergeMap';
 export { min } from './min';
